Refactorisations
 - Rename CudaContextSingleton as CudaContext, and move it into NiftyReg namespace
 - Rename NiftyReg_CudaBlock100 as BlockSize, and move it into NiftyReg namespace
 - Move BlockSize implementation into the header
 - Change the type of BlockSize members as unsigned
 - Move BlockSize instance into CudaContext
 - Use unsigned instead of size_t in CUDA kernels
 - Initialise the CUDA or OpenCL device in Platform's constructor
 - Rename `unsigned int`s as `unsigned`

+++ reg-lib/cuda/_reg_optimiser_kernels.cu
@@ -1,8 +1,8 @@
 /* *************************************************************** */
 __global__ void reg_initialiseConjugateGradient_kernel(float4 *conjugateGCuda,
                                                        cudaTextureObject_t gradientImageTexture,
-                                                       const size_t nVoxels) {
-    const size_t tid = ((size_t)blockIdx.y * gridDim.x + blockIdx.x) * blockDim.x + threadIdx.x;
+                                                       const unsigned nVoxels) {
+    const unsigned tid = (blockIdx.y * gridDim.x + blockIdx.x) * blockDim.x + threadIdx.x;
     if (tid < nVoxels) {
         const float4 gradValue = tex1Dfetch<float4>(gradientImageTexture, tid);
         conjugateGCuda[tid] = make_float4(-gradValue.x, -gradValue.y, -gradValue.z, 0);
@@ -13,8 +13,8 @@ __global__ void reg_GetConjugateGradient1_kernel(float2 *sums,
                                                  cudaTextureObject_t gradientImageTexture,
                                                  cudaTextureObject_t conjugateGTexture,
                                                  cudaTextureObject_t conjugateHTexture,
-                                                 const size_t nVoxels) {
-    const size_t tid = ((size_t)blockIdx.y * gridDim.x + blockIdx.x) * blockDim.x + threadIdx.x;
+                                                 const unsigned nVoxels) {
+    const unsigned tid = (blockIdx.y * gridDim.x + blockIdx.x) * blockDim.x + threadIdx.x;
     if (tid < nVoxels) {
         const float4 valueH = tex1Dfetch<float4>(conjugateHTexture, tid);
         const float4 valueG = tex1Dfetch<float4>(conjugateGTexture, tid);
@@ -30,9 +30,9 @@ __global__ void reg_GetConjugateGradient1_kernel(float2 *sums,
 __global__ void reg_GetConjugateGradient2_kernel(float4 *gradientImageCuda,
                                                  float4 *conjugateGCuda,
                                                  float4 *conjugateHCuda,
-                                                 const size_t nVoxels,
+                                                 const unsigned nVoxels,
                                                  const float scale) {
-    const size_t tid = ((size_t)blockIdx.y * gridDim.x + blockIdx.x) * blockDim.x + threadIdx.x;
+    const unsigned tid = (blockIdx.y * gridDim.x + blockIdx.x) * blockDim.x + threadIdx.x;
     if (tid < nVoxels) {
         // G = - grad
         float4 gradGValue = gradientImageCuda[tid];
@@ -54,12 +54,12 @@ __global__ void reg_GetConjugateGradient2_kernel(float4 *gradientImageCuda,
 __global__ void reg_updateControlPointPosition_kernel(float4 *controlPointImageCuda,
                                                       cudaTextureObject_t bestControlPointTexture,
                                                       cudaTextureObject_t gradientImageTexture,
-                                                      const size_t nVoxels,
+                                                      const unsigned nVoxels,
                                                       const float scale,
                                                       const bool optimiseX,
                                                       const bool optimiseY,
                                                       const bool optimiseZ) {
-    const size_t tid = ((size_t)blockIdx.y * gridDim.x + blockIdx.x) * blockDim.x + threadIdx.x;
+    const unsigned tid = (blockIdx.y * gridDim.x + blockIdx.x) * blockDim.x + threadIdx.x;
     if (tid < nVoxels) {
         float4 value = controlPointImageCuda[tid];
         const float4 bestValue = tex1Dfetch<float4>(bestControlPointTexture, tid);
